requirements.txt
# Core dependencies
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web scraping and search
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
lxml>=4.9.0

# CLI
rich>=13.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
responses>=0.23.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

